upgrade database without losing user data
carefully update onUpgrade() when bumping database versions to make
sure user data is saved.

+++ src/com/seafile/seadroid2/account/AccountDBHelper.java
@@ -2,6 +2,7 @@ package com.seafile.seadroid2.account;
 
 import java.util.List;
 
+import android.util.Log;
 import com.google.common.collect.Lists;
 import com.seafile.seadroid2.data.ServerInfo;
 import com.seafile.seadroid2.provider.AccountNotifier;
@@ -16,8 +17,21 @@ import android.database.sqlite.SQLiteOpenHelper;
  * A helper class to manage {@link #DATABASE_NAME} database creation and version management.
  */
 public class AccountDBHelper extends SQLiteOpenHelper {
+    public static final String DEBUG_TAG = "AccountDBHelper";
+
+    // NOTE: carefully update onUpgrade() when bumping database versions to make
+    // sure user data is saved.
+
+    /** version of adding a new table @{link #ACCOUNT_TABLE_NAME} */
+    private static final int VER_ADD_NEW_TABLE_ACCOUNT = 1;
+    /** version of adding a new table @{link #SERVER_INFO_TABLE_NAME} */
+    private static final int VER_ADD_NEW_TABLE_SERVER_INFO = 2;
+    /** version of adding a new table @{link #SERVER_INFO_TABLE_NAME} without losing data */
+    private static final int VER_ADD_NEW_TABLE_SERVER_INFO_MIGRATION = 3;
+
     // If you change the database schema, you must increment the database version.
-    private static final int DATABASE_VERSION = 2;
+    private static final int DATABASE_VERSION = VER_ADD_NEW_TABLE_SERVER_INFO_MIGRATION;
+
     private static final String DATABASE_NAME = "account.db";
 
     private static final String ACCOUNT_TABLE_NAME = "Account";
@@ -69,11 +83,42 @@ public class AccountDBHelper extends SQLiteOpenHelper {
     }
 
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        // This database is only a cache for online data, so its upgrade policy is
-        // to simply to discard the data and start over
-        db.execSQL("DROP TABLE IF EXISTS " + ACCOUNT_TABLE_NAME + ";");
-        db.execSQL("DROP TABLE IF EXISTS " + SERVER_INFO_TABLE_NAME + ";");
-        onCreate(db);
+        Log.d(DEBUG_TAG, "onUpgrade() from " + oldVersion + " to " + newVersion);
+
+        // NOTE: This switch statement is designed to handle cascading database
+        // updates, starting at the current version and falling through to all
+        // future upgrade cases. Only use "break;" when you want to drop and
+        // recreate the entire database.
+
+        // Current DB version. We update this variable as we perform upgrades to reflect
+        // the current version we are in.
+        int version = oldVersion;
+
+        switch (version) {
+            case VER_ADD_NEW_TABLE_ACCOUNT:
+                // Version 2 added new table "ServerInfo"
+                db.execSQL("DROP TABLE IF EXISTS " + SERVER_INFO_TABLE_NAME + ";");
+                db.execSQL(SQL_CREATE_SERVER_INFO_TABLE);
+                version = VER_ADD_NEW_TABLE_SERVER_INFO;
+
+            case VER_ADD_NEW_TABLE_SERVER_INFO:
+                // Version 3 added new table "ServerInfo" without losing user data
+                db.execSQL("DROP TABLE IF EXISTS " + SERVER_INFO_TABLE_NAME + ";");
+                db.execSQL(SQL_CREATE_SERVER_INFO_TABLE);
+                version = VER_ADD_NEW_TABLE_SERVER_INFO_MIGRATION;
+        }
+
+        Log.d(DEBUG_TAG, "after upgrade logic, at version " + version);
+
+        // at this point, we ran out of upgrade logic, so if we are still at the wrong
+        // version, we have no choice but to delete everything and create everything again.
+        if (version != DATABASE_VERSION) {
+            Log.w(DEBUG_TAG, "Destroying old data during upgrade");
+            db.execSQL("DROP TABLE IF EXISTS " + ACCOUNT_TABLE_NAME + ";");
+            db.execSQL("DROP TABLE IF EXISTS " + SERVER_INFO_TABLE_NAME + ";");
+            onCreate(db);
+            version = DATABASE_VERSION;
+        }
     }
 
     public void onDowngrade(SQLiteDatabase db, int oldVersion, int newVersion) {
